Boolean Query Processing Based on Postings List
In this Project, postings list from the RCV1 news corpus has been used
(http://www.daviddlewis.com/resources/testcollections/rcv1/). This
project implements modules that return documents based on
term-at-a-time with the postings list ordered by term frequencies, and
document-at-a-time with the postings list ordered by doc IDs for a set
of queries.